Test 클래스 변경(SsdTest -> CommanderTest)
